Refactor Poll. Move PollAnswer into poll.py and rename it to PollOption
